replace rote recommendation
Anon suggested: "I'm the one who originally wrote this, but I no longer
believe it based on my own experience. While this deck appeared to me in
the beginning like it would be the best resource for rote kanji study on
the grounds of how complete and comprehensive it is, I have come to
realise that it has a major problem in that a lot of the kanji in it
have an absurd amount of meanings that you can't possibly hope to
memorise.

If I were to give advice to someone now on choosing a deck to study
kanji via the rote method, I would tell them to pick one of the mnemonic
decks and just strip the mnemonics from it, the reason being that the
set of meanings they present for any given kanji are much more concise
(only 1 in the case of the RTK and KD decks, which might be overly
concise, and up to 3 in the KKLC deck) and thus much easier to remember.

I recommend perhaps deleting this entry."